+++ packages/elasticsearch/src/example-types/metricGroups/numberRangesGroupStats.js
@@ -11,7 +11,7 @@ let buildQuery = ({ groupField: field, statsField, stats, ranges }) => ({
 
 module.exports = {
   buildQuery,
-  validContext: node => node.groupField && node.statsField && node.ranges,
+  validContext: node => node.groupField && node.ranges,
   async result(node, search) {
     let response = await search(buildQuery(node))
     return { results: simplifyBuckets(response.aggregations.groups.buckets) }